mythbackend: rename commandlineparser.* to mythbackend_...
